Implement terminal init and cleanup
